Rename: 이미지 폴더 일괄 이동
Vercel 환경에서의 이미지 깨짐 문제를 수정하기 위해 이미지 폴더를 public으로 이동하였습니다.
