+++ frontend/public/images/foodlier_logo.png
