+++ frontend/src/components/chef/ChefData.ts
@@ -8,17 +8,17 @@ export const chefData: Chef[] = [
   {
     chefId: 1,
     nickname: '닉네임1',
-    profileUrl: '/src/assets/chef.svg',
+    profileUrl: '../../../public/images/chef.svg',
   },
   {
     chefId: 2,
     nickname: '닉네임2',
-    profileUrl: '/src/assets/chef.svg',
+    profileUrl: '../../../public/images/chef.svg',
   },
   {
     chefId: 3,
     nickname: '닉네임3',
-    profileUrl: '/src/assets/chef.svg',
+    profileUrl: '../../../public/images/chef.svg',
   },
 ]
