+++ frontend/src/pages/auth/LoginPage.tsx
@@ -1,9 +1,9 @@
 import { useNavigate } from 'react-router-dom'
 import axios from 'axios'
 import { useState, useEffect } from 'react'
-import logo from '../../assets/foodlier_logo.png'
-import naver from '../../assets/btn_naver.png'
-import kakao from '../../assets/btn_kakao.png'
+import kakao from '../../../public/images/auths/btn_kakao.png'
+import logo from '../../../public/images/foodlier_logo.png'
+import naver from '../../../public/images/auths/btn_naver.png'
 import * as S from '../../styles/auth/LoginPage.styled'
 
 interface ResType {
